chore: address PR comments, and improve some names of props
